fix(vst3): cast state activateBus i16 → u8 (TBool VST3)
activateBus signature : state: TBool = u8 (pas i16 comme supposé). Sans
le cast explicite, l'inférence Rust ne devine pas le type quand
est extrait dans une variable intermédiaire avant le call.

+++ jamodio-vst3-host/src/host.rs
@@ -288,7 +288,7 @@ impl Instance {
                     self.component.getBusCount(media as i32, dir as i32)
                 };
                 for idx in 0..n {
-                    let state: i16 = if idx == 0 { 1 } else { 0 };
+                    let state: u8 = if idx == 0 { 1 } else { 0 };
                     let act_ok = unsafe {
                         self.component.activateBus(media as i32, dir as i32, idx, state)
                     };
